インフラ層で扱う型の見直し
