+++ functions/src/usecases/contracts/ddb-operations.ts
@@ -1,16 +1,20 @@
-import { TaskItem } from '../../domain/taskItem';
-import { CreateTaskRequest } from '../../handlers/http/requestSchemas/task-requests';
+import { TaskItem } from '../../infrastructure/ddb/schemas/taskItem';
 
 export type DdbOperation<T, P extends unknown[]> = (...args: P) => Promise<T>;
 
-export type CreateTaskItem = DdbOperation<string, [CreateTaskRequest]>;
+export type CreateTaskItem = DdbOperation<string, [CreateTaskPayload]>;
 export type GetTaskItem = DdbOperation<TaskItem | null, [string]>;
 export type UpdateTaskItem = DdbOperation<
   TaskItem,
   [string, TaskUpdateAtLeastOne]
 >;
 
-type TaskUpdateData = {
+export type CreateTaskPayload = {
+  title: string;
+  description?: string;
+};
+
+type UpdateTaskPayload = {
   title: string;
   description: string;
 };
@@ -18,4 +22,4 @@ type TaskUpdateData = {
 type AtLeastOne<T, U = { [K in keyof T]: Pick<T, K> }> = Partial<T> &
   U[keyof U];
 
-export type TaskUpdateAtLeastOne = AtLeastOne<TaskUpdateData>;
+export type TaskUpdateAtLeastOne = AtLeastOne<UpdateTaskPayload>;
