+++ functions/tests/medium/infrastructure/ddb/tasks-table.test.ts
@@ -1,4 +1,4 @@
-import { TaskItem } from '../../../../src/domain/taskItem';
+import { TaskItem } from '../../../../src/infrastructure/ddb/schemas/taskItem';
 import { CreateTaskRequest } from '../../../../src/handlers/http/requestSchemas/task-requests';
 
 import {
